+++ Enzymes/gin_infograph.py
@@ -77,7 +77,7 @@ class Encoder(torch.nn.Module):
                 data.to(device)
                 x, edge_index, batch = data.x, data.edge_index, data.batch
 
-                x_graph, x_node = self.forward(x[:,:18], edge_index, batch)
+                x_graph, x_node = self.forward(x[:,:18].double(), edge_index, batch)
 
 
                 ret_node.append(x_node.cpu().numpy())
